float type to double

+++ car/Car/CarMovement.cpp
@@ -10,21 +10,21 @@
 			int pin_left_motor_back,
 			int pin_right_motor_go,
 			int pin_right_motor_back,
-			float car_width) :
+			double car_width) :
 			pin_left_motor_go(pin_left_motor_go),
 			pin_left_motor_back(pin_left_motor_back),
 			pin_right_motor_go(pin_right_motor_go),
 			pin_right_motor_back(pin_right_motor_back),
 			car_width(car_width) {}
 
-	float CarMovement::getAngularSpeed(const float &average_speed, const float &radius){
+	double CarMovement::getAngularSpeed(const double &average_speed, const double &radius){
 		return 2 * average_speed / (2 * radius + car_width); // 각속도
 	}
 
 
-	void CarMovement::turnSpecificDegree(const float &angular_speed, const float &degree) {
-		float angle_radian = degree * PI / 180;
-		float delay_time = angle_radian / angular_speed;
+	void CarMovement::turnSpecificDegree(const double &angular_speed, const double &degree) {
+		double angle_radian = degree * PI / 180;
+		double delay_time = angle_radian / angular_speed;
 
 		delay(delay_time * 1000);
 		stop();
@@ -32,7 +32,7 @@
 
 
 	// 앞으로, 뒤로
-	void CarMovement::goForward	(const float &speed, float distance){
+	void CarMovement::goForward	(const double &speed, double distance){
 		setMotorSpeed(speed, speed);
 		if (distance) {
 			delay(speed / distance * 1000);
@@ -40,7 +40,7 @@
 		}
 	}
 
-	void CarMovement::goBackward	(const float &speed, float distance) {
+	void CarMovement::goBackward	(const double &speed, double distance) {
 		setMotorSpeed(-speed, -speed);
 		if (distance) {
 			delay(speed / distance * 1000);
@@ -55,10 +55,10 @@
 
 	// 회전
 	// '라디안'이 아니라 '도'로
-	void CarMovement::turnLeft	(const float &average_speed, const float &radius, const float &degree= 0) {
-		float angular_speed = getAngularSpeed(average_speed, radius);
-		float right_speed = angular_speed * (radius + car_width);
-		float left_speed = angular_speed * radius;
+	void CarMovement::turnLeft	(const double &average_speed, const double &radius, const double &degree= 0) {
+		double angular_speed = getAngularSpeed(average_speed, radius);
+		double right_speed = angular_speed * (radius + car_width);
+		double left_speed = angular_speed * radius;
 
 		setMotorSpeed(right_speed, left_speed);
 
@@ -67,10 +67,10 @@
 		}
 	}
 
-	void CarMovement::turnRight	(const float &average_speed, const float &radius, const float &degree = 0){
-		float angular_speed = getAngularSpeed(average_speed, radius);
-		float right_speed = angular_speed * radius;
-		float left_speed = angular_speed * (radius + car_width);
+	void CarMovement::turnRight	(const double &average_speed, const double &radius, const double &degree = 0){
+		double angular_speed = getAngularSpeed(average_speed, radius);
+		double right_speed = angular_speed * radius;
+		double left_speed = angular_speed * (radius + car_width);
 
 		setMotorSpeed(right_speed, left_speed);
 
@@ -81,10 +81,10 @@
 
 
 	// 제자리리에서 회전
-	void CarMovement::spinLeft	(const float &average_speed, const float &radius, const float &degree = 0){
-		float angular_speed = getAngularSpeed(average_speed, radius);
-		float right_speed = angular_speed * (radius - car_width);
-		float left_speed = angular_speed * (radius + car_width);
+	void CarMovement::spinLeft	(const double &average_speed, const double &radius, const double &degree = 0){
+		double angular_speed = getAngularSpeed(average_speed, radius);
+		double right_speed = angular_speed * (radius - car_width);
+		double left_speed = angular_speed * (radius + car_width);
 
 		setMotorSpeed(right_speed, left_speed);
 
@@ -93,10 +93,10 @@
 		}
 	}
 
-	void CarMovement::spinRight	(const float &average_speed, const float &radius, const float &degree = 0){
-		float angular_speed = getAngularSpeed(average_speed, radius);
-		float right_speed = angular_speed * (radius + car_width);
-		float left_speed = angular_speed * (radius - car_width);
+	void CarMovement::spinRight	(const double &average_speed, const double &radius, const double &degree = 0){
+		double angular_speed = getAngularSpeed(average_speed, radius);
+		double right_speed = angular_speed * (radius + car_width);
+		double left_speed = angular_speed * (radius - car_width);
 
 		setMotorSpeed(right_speed, left_speed);
 
@@ -106,7 +106,7 @@
 	}
 
 	//CarMoveAnalog
-	void CarMovementAnalog::setMotorSpeed(const float &right_speed, const float &left_speed) {
+	void CarMovementAnalog::setMotorSpeed(const double &right_speed, const double &left_speed) {
 		if (right_speed  >= 0) {
 			analogWrite(pin_right_motor_go, right_speed);
 			analogWrite(pin_right_motor_back, 0);
@@ -124,7 +124,7 @@
 	}
 
 	// CarMovementDigital
-	void CarMovementDigital::setMotorSpeed(const float &right_speed, const float &left_speed) {
+	void CarMovementDigital::setMotorSpeed(const double &right_speed, const double &left_speed) {
 		if (right_speed  >= 0) {
 			digitalWrite(pin_right_motor_go, HIGH);
 			digitalWrite(pin_right_motor_back, LOW);
